moves test data generation files into folder
