+++ api/test/generate-data/factories/group_factory.js
@@ -1,6 +1,6 @@
 const factory = require('factory-girl').factory;
 const factory_helper = require('./factory_helper');
-const Group = require('../../helpers/models/group');
+const Group = require('../../../helpers/models/group');
 let faker = require('faker/locale/en');
 
 const factoryName = Group.modelName;
